Integration with ICMM
The indicator processes now have one single parameter: the URL pointing to the relevant ICMM worldstate.
The process asks ICMM where to find the actual data (only OOI-WSR so long)
The indicator value is calculated and stored in OOI-WSR
The reference to the value is stored in the ICMM worldstate

The test-page has also been updated.

The OrionListener.py still needs an update. Will be done as soon as ICMM with PubSub is in use.